requirements.txt
absl-py==1.4.0
aiofiles==25.1.0
alphafold-colabfold==2.3.6
appdirs==1.4.4
astunparse==1.6.3
attrs==23.2.0
Automat==22.10.0
Babel==2.10.3
bcrypt==3.2.2
biopython==1.82
blinker==1.7.0
certifi==2023.11.17
chardet==5.2.0
chex==0.1.90
click==8.1.6
cloud-init==25.1.4
colabfold==1.5.4
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
constantly==23.10.4
contourpy==1.3.3
cryptography==41.0.7
cycler==0.12.1
dbus-python==1.3.2
distro==1.9.0
distro-info==1.7+build1
dm-haiku==0.0.11
dm-tree==0.1.9
etils==1.13.0
filelock==3.20.3
flatbuffers==25.12.19
flax==0.10.4
fonttools==4.61.1
fsspec==2026.1.0
gast==0.7.0
google-pasta==0.2.0
grpcio==1.76.0
h5py==3.15.1
httplib2==0.20.4
humanize==4.15.0
hyperlink==21.0.0
idna==3.6
immutabledict==4.2.2
importlib-metadata==4.13.0
importlib_resources==6.5.2
incremental==22.10.0
jax==0.4.35
jaxlib==0.4.35
Jinja2==3.1.2
jmp==0.0.4
joblib==1.5.3
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
keras==3.13.1
kiwisolver==1.4.9
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
libclang==18.1.1
Markdown==3.10
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.8
mdurl==0.1.2
ml_collections==1.1.0
ml_dtypes==0.5.4
mpmath==1.3.0
msgpack==1.1.2
namex==0.1.0
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.6.1
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
oauthlib==3.2.2
opt_einsum==3.4.0
optax==0.2.5
optree==0.18.0
orbax-checkpoint==0.11.5
packaging==25.0
pandas==1.5.3
pillow==12.1.0
protobuf==6.33.4
psutil==7.2.1
py3Dmol==2.5.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycurl==7.45.3
Pygments==2.17.2
PyGObject==3.48.2
PyHamcrest==2.1.0
PyJWT==2.7.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu5.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
scikit-learn==1.8.0
scipy==1.17.0
service-identity==24.1.0
setuptools==68.1.2
simplejson==3.20.2
six==1.16.0
sympy==1.14.0
systemd-python==235
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
tensorflow_cpu==2.20.0
tensorstore==0.1.80
termcolor==3.3.0
threadpoolctl==3.6.0
toolz==1.1.0
torch==2.9.1
tqdm==4.67.1
treescope==0.1.10
triton==3.5.1
Twisted==24.3.0
typing_extensions==4.15.0
ubuntu-pro-client==8001
unattended-upgrades==0.1
urllib3==2.0.7
wadllib==1.3.6
Werkzeug==3.1.5
wheel==0.42.0
wrapt==2.0.1
zipp==3.23.0
zope.interface==6.1